moved Stuff which is only supported sinc gtk3.8 to seperate file
